@@ -25,16 +25,16 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^17.3.9",
-    "@angular/cdk": "^17.3.9",
-    "@angular/common": "^17.3.9",
-    "@angular/compiler": "^17.3.9",
-    "@angular/core": "^17.3.9",
-    "@angular/forms": "^17.3.9",
-    "@angular/material": "^17.3.9",
-    "@angular/platform-browser": "^17.3.9",
-    "@angular/platform-browser-dynamic": "^17.3.9",
-    "@angular/router": "^17.3.9",
+    "@angular/animations": "^18.0.0",
+    "@angular/cdk": "^18.0.0",
+    "@angular/common": "^18.0.0",
+    "@angular/compiler": "^18.0.0",
+    "@angular/core": "^18.0.0",
+    "@angular/forms": "^18.0.0",
+    "@angular/material": "^18.0.0",
+    "@angular/platform-browser": "^18.0.0",
+    "@angular/platform-browser-dynamic": "^18.0.0",
+    "@angular/router": "^18.0.0",
     "@auth0/angular-jwt": "^5.2.0",
     "@cisstech/nge": "^17.7.1",
     "@myrmidon/auth-jwt-admin": "4.1.0",
@@ -50,7 +50,7 @@
     "@myrmidon/cadmus-refs-viaf-lookup": "^5.0.0",
     "@myrmidon/cadmus-text-block-view": "^5.0.0",
     "@myrmidon/cadmus-text-ed": "^5.0.1",
-    "@myrmidon/cadmus-text-ed-md": "^5.0.2",
+    "@myrmidon/cadmus-text-ed-md": "^5.0.3",
     "@myrmidon/cadmus-text-ed-txt": "^5.0.0",
     "@myrmidon/ng-mat-tools": "3.0.0",
     "@myrmidon/ng-tools": "3.0.0",
@@ -66,13 +66,13 @@
     "monaco-editor": "^0.48.0",
     "rxjs": "~7.8.1",
     "tslib": "^2.6.2",
-    "zone.js": "~0.14.5"
+    "zone.js": "~0.14.6"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^17.3.7",
-    "@angular/cli": "^17.3.7",
-    "@angular/compiler-cli": "^17.3.9",
-    "@angular/localize": "^17.3.9",
+    "@angular-devkit/build-angular": "^18.0.0",
+    "@angular/cli": "^18.0.0",
+    "@angular/compiler-cli": "^18.0.0",
+    "@angular/localize": "^18.0.0",
     "@types/d3-dispatch": "^3.0.6",
     "@types/d3-drag": "^3.0.7",
     "@types/d3-timer": "^3.0.2",
@@ -84,7 +84,7 @@
     "karma-coverage": "~2.2.1",
     "karma-jasmine": "~5.1.0",
     "karma-jasmine-html-reporter": "~2.1.0",
-    "ng-packagr": "^17.3.0",
+    "ng-packagr": "^18.0.0",
     "typescript": "~5.4.5"
   }
 }
