@@ -1,5 +1,5 @@
 import { TestBed } from '@angular/core/testing';
-import { HttpClient, HttpClientModule } from '@angular/common/http';
+import { HttpClient, provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 
 import { EnvServiceProvider } from '@myrmidon/ng-tools';
 
@@ -8,9 +8,9 @@ import { FacetService } from './facet.service';
 describe('FacetService', () => {
   beforeEach(() => {
     TestBed.configureTestingModule({
-      imports: [HttpClientModule],
-      providers: [HttpClient, EnvServiceProvider],
-    });
+    imports: [],
+    providers: [HttpClient, EnvServiceProvider, provideHttpClient(withInterceptorsFromDi())]
+});
   });
 
   it('should be created', () => {
