@@ -1,6 +1,6 @@
 import { TestBed } from '@angular/core/testing';
 
-import { HttpClientModule, HttpClient } from '@angular/common/http';
+import { HttpClient, provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 import { CadmusCoreModule } from '@myrmidon/cadmus-core';
 import { EnvServiceProvider } from '@myrmidon/ng-tools';
 
@@ -9,9 +9,9 @@ import { ItemBrowserService } from './item-browser.service';
 describe('ItemBrowserService', () => {
   beforeEach(() =>
     TestBed.configureTestingModule({
-      imports: [HttpClientModule, CadmusCoreModule],
-      providers: [HttpClient, EnvServiceProvider],
-    })
+    imports: [CadmusCoreModule],
+    providers: [HttpClient, EnvServiceProvider, provideHttpClient(withInterceptorsFromDi())]
+})
   );
   it('should be created', () => {
     const service: ItemBrowserService = TestBed.inject(ItemBrowserService);
