@@ -1,5 +1,5 @@
 import { TestBed } from '@angular/core/testing';
-import { HttpClientModule, HttpClient } from '@angular/common/http';
+import { HttpClient, provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 
 import { ItemService } from './item.service';
 import { EnvServiceProvider } from '@myrmidon/ng-tools';
@@ -7,9 +7,9 @@ import { EnvServiceProvider } from '@myrmidon/ng-tools';
 describe('ItemService', () => {
   beforeEach(() =>
     TestBed.configureTestingModule({
-      imports: [HttpClientModule],
-      providers: [HttpClient, EnvServiceProvider],
-    })
+    imports: [],
+    providers: [HttpClient, EnvServiceProvider, provideHttpClient(withInterceptorsFromDi())]
+})
   );
 
   it('should be created', () => {
