@@ -1,15 +1,15 @@
 import { TestBed } from '@angular/core/testing';
 
 import { OperatorGuardService } from './operator-guard.service';
-import { HttpClientModule, HttpClient } from '@angular/common/http';
+import { HttpClient, provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 import { RouterTestingModule } from '@angular/router/testing';
 
 describe('OperatorGuardService', () => {
   beforeEach(() =>
     TestBed.configureTestingModule({
-      imports: [HttpClientModule, RouterTestingModule],
-      providers: [HttpClient, RouterTestingModule],
-    })
+    imports: [RouterTestingModule],
+    providers: [HttpClient, RouterTestingModule, provideHttpClient(withInterceptorsFromDi())]
+})
   );
 
   it('should be created', () => {
