@@ -1,18 +1,19 @@
 import { TestBed } from '@angular/core/testing';
 
 import { ThesaurusService } from './thesaurus.service';
-import { HttpClientModule, HttpClient } from '@angular/common/http';
+import { HttpClient, provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 
 xdescribe('TagService', () => {
   beforeEach(() => {
     TestBed.configureTestingModule({
-      imports: [HttpClientModule],
-      providers: [
+    imports: [],
+    providers: [
         HttpClient,
         { provide: 'apiEndpoint', useValue: 'none' },
         { provide: 'databaseId', useValue: 'cadmus' },
-      ],
-    });
+        provideHttpClient(withInterceptorsFromDi()),
+    ]
+});
   });
 
   it('should be created', () => {
