@@ -1,17 +1,18 @@
 import { TestBed } from '@angular/core/testing';
 import { UserService } from './user.service';
-import { HttpClientModule, HttpClient } from '@angular/common/http';
+import { HttpClient, provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 
 xdescribe('Service: User', () => {
   beforeEach(() => {
     TestBed.configureTestingModule({
-      imports: [HttpClientModule],
-      providers: [
+    imports: [],
+    providers: [
         HttpClient,
         { provide: 'apiEndpoint', useValue: 'none' },
         { provide: 'databaseId', useValue: 'cadmus' },
-      ],
-    });
+        provideHttpClient(withInterceptorsFromDi()),
+    ]
+});
   });
 
   it('should be created', () => {
