@@ -1,14 +1,11 @@
 import { CommonModule } from '@angular/common';
-import { HttpClientModule } from '@angular/common/http';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 import { NgModule } from '@angular/core';
 
 import { NgToolsModule } from '@myrmidon/ng-tools';
 
 import { SortPipe } from './pipes/sort.pipe';
 
-@NgModule({
-  imports: [CommonModule, HttpClientModule, NgToolsModule],
-  declarations: [SortPipe],
-  exports: [SortPipe],
-})
+@NgModule({ declarations: [SortPipe],
+    exports: [SortPipe], imports: [CommonModule, NgToolsModule], providers: [provideHttpClient(withInterceptorsFromDi())] })
 export class CadmusCoreModule {}
