@@ -1,5 +1,5 @@
 import { CommonModule } from '@angular/common';
-import { HttpClientModule } from '@angular/common/http';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 import { NgModule } from '@angular/core';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 
@@ -33,51 +33,45 @@ import { LinkedLiteralFilterComponent } from './components/linked-literal-filter
 import { LinkedNodeFilterComponent } from './components/linked-node-filter/linked-node-filter.component';
 import { GraphNodeLabelPipe } from './pipes/graph-node-label.pipe';
 
-@NgModule({
-  declarations: [
-    GraphNodeLabelPipe,
-    GraphWalkerComponent,
-    LinkedLiteralFilterComponent,
-    LinkedNodeFilterComponent,
-    TripleFilterComponent,
-  ],
-  imports: [
-    CommonModule,
-    FormsModule,
-    ReactiveFormsModule,
-    HttpClientModule,
-    // material
-    MatAutocompleteModule,
-    MatButtonModule,
-    MatCheckboxModule,
-    MatChipsModule,
-    MatFormFieldModule,
-    MatIconModule,
-    MatInputModule,
-    MatListModule,
-    MatPaginatorModule,
-    MatProgressBarModule,
-    MatSelectModule,
-    MatTabsModule,
-    MatTooltipModule,
-    // vendor
-    NgxGraphModule,
-    // Cadmus
-    NgToolsModule,
-    NgMatToolsModule,
-    CadmusCoreModule,
-    CadmusApiModule,
-    // CadmusRefsLookupModule,
-    RefLookupComponent,
-    CadmusGraphUiModule
-  ],
-  exports: [
-    GraphNodeLabelPipe,
-    GraphWalkerComponent,
-    CadmusApiModule,
-    LinkedLiteralFilterComponent,
-    LinkedNodeFilterComponent,
-    TripleFilterComponent,
-  ],
-})
+@NgModule({ declarations: [
+        GraphNodeLabelPipe,
+        GraphWalkerComponent,
+        LinkedLiteralFilterComponent,
+        LinkedNodeFilterComponent,
+        TripleFilterComponent,
+    ],
+    exports: [
+        GraphNodeLabelPipe,
+        GraphWalkerComponent,
+        CadmusApiModule,
+        LinkedLiteralFilterComponent,
+        LinkedNodeFilterComponent,
+        TripleFilterComponent,
+    ], imports: [CommonModule,
+        FormsModule,
+        ReactiveFormsModule,
+        // material
+        MatAutocompleteModule,
+        MatButtonModule,
+        MatCheckboxModule,
+        MatChipsModule,
+        MatFormFieldModule,
+        MatIconModule,
+        MatInputModule,
+        MatListModule,
+        MatPaginatorModule,
+        MatProgressBarModule,
+        MatSelectModule,
+        MatTabsModule,
+        MatTooltipModule,
+        // vendor
+        NgxGraphModule,
+        // Cadmus
+        NgToolsModule,
+        NgMatToolsModule,
+        CadmusCoreModule,
+        CadmusApiModule,
+        // CadmusRefsLookupModule,
+        RefLookupComponent,
+        CadmusGraphUiModule], providers: [provideHttpClient(withInterceptorsFromDi())] })
 export class CadmusGraphUiExModule {}
